// src/FE/src/api/generated/models/jobListItemViewModel.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Workslip.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { AssignedUserResponse } from './assignedUserResponse';
import type { CustomerViewModel } from './customerViewModel';
import type { JobStatus } from './jobStatus';

export interface JobListItemViewModel {
  id: string;
  organizationId: string;
  customer: null | CustomerViewModel;
  /** @nullable */
  reportNumber: string | null;
  status: JobStatus;
  installationTypes: string[];
  assignedUsers: AssignedUserResponse[];
  softDeleted: boolean;
  /**
     * @nullable
     * @pattern ^-?(?:0|[1-9]\d*)(?:\.\d+)?$
     */
  totalHours: number | string | null;
  updatedAt: string;
  /** @nullable */
  reportDate: string | null;
  jobType: string;
  /** @nullable */
  destinationAddress: string | null;
  /** @nullable */
  taskDescription: string | null;
  isSeen: boolean;
  isNewRejection: boolean;
}